update cart payout

+++ app/src/main/java/com/example/doankotlin/Domain/CartItems.kt
@@ -2,7 +2,7 @@ package com.example.doankotlin.Domain
 
 data class CartItems(
     var Title: String?= null,
-    var Price: Double?= null,
+    var Price: Int?= null,
     var ImagePath: String?= null,
     var Quantity: Int? = null,
 )
