+++ app/src/main/java/com/example/doankotlin/Domain/Foods.java
@@ -8,7 +8,7 @@ public class Foods implements Serializable {
     private boolean BestFood;
     private int Id;
     private int LocationId;
-    private double Price;
+    private int Price;
     private String ImagePath;
     private int PriceId;
     private double Star;
@@ -65,11 +65,11 @@ public class Foods implements Serializable {
         LocationId = locationId;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return Price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         Price = price;
     }
